doc/*.ksy: Fix compilation errors
The ternary/conditional operator '?' includes a ':', which is part of
the YAML key/value syntax. Depending on the YAML parser, this in turn
may produce parsing errors. Apparently they switched the YAML parser on
the web IDE, so now all these YAML files didn't parse.
In the documentation they do point this out as a compatibility concern
and suggest to quote such expressions.
https://userunknownfactor.github.io/user_guide.html#_operators

Additionally, the new YAML parser pointed out duplicate keys, which I
also fixed.